compileEnum: support checkExists

+++ testdata/helloprj/hello.c
@@ -5,6 +5,6 @@
 int main() {
     const int N = 32;
     char msg[N] = "Hello";
-    printf("%d %s, %s\n", foo(), msg, bar());
+    printf("%d %s, %s\n", (int)foo(), msg, bar());
     return 0;
 }
