+++ testdata/libc/src/foo.h
@@ -1,3 +1,5 @@
-typedef int foo_t;
+#include <stdlib.h>
+
+typedef size_t foo_t;
 
 foo_t foo();
